Refactored deployments pages
