+++ common/src/main/java/org/jboss/hal/testsuite/page/runtime/DomainDeploymentEntryPoint.java
@@ -1,0 +1,8 @@
+package org.jboss.hal.testsuite.page.runtime;
+
+import org.jboss.arquillian.graphene.page.Location;
+import org.jboss.hal.testsuite.page.BasePage;
+
+@Location("#domain-deployments")
+public class DomainDeploymentEntryPoint extends BasePage {
+}
