+++ common/src/main/java/org/jboss/hal/testsuite/page/runtime/StandaloneDeploymentEntryPoint.java
@@ -1,0 +1,8 @@
+package org.jboss.hal.testsuite.page.runtime;
+
+import org.jboss.arquillian.graphene.page.Location;
+import org.jboss.hal.testsuite.page.BasePage;
+
+@Location("#standalone-deployments")
+public class StandaloneDeploymentEntryPoint extends BasePage {
+}
